Make onEach and on inline functions
They will be called a lot in different screens, it seems worth having
classes saved. Although "val transitions: MutableList" will be exposed
as part of the ABI, but it shouldn't be used by anyone sane, its on the
DslObject. And is hidden from IDE.

Though when it changes it will break binary compatibility.